Mac build fix
Review URL: http://codereview.chromium.org/16610


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@8136 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/waitable_event_posix.cc
@@ -303,7 +303,7 @@ size_t WaitableEvent::WaitMany(WaitableEvent** raw_waitables,
 // -----------------------------------------------------------------------------
 // static
 unsigned WaitableEvent::EnqueueMany
-    (std::pair<WaitableEvent*, unsigned>* waitables,
+    (std::pair<WaitableEvent*, size_t>* waitables,
      unsigned count, Waiter* waiter) {
   if (!count)
     return 0;
